Revert "Refactor logic into separate project (#61)"
This reverts commit ac119ea1531ed1ce344ff35cc0ef06e0ee5acd80.
