--- src/Valleysoft.Dredge.Core/DockerfileOptions.cs
@@ -1,7 +1,0 @@
-﻿namespace Valleysoft.Dredge.Core;
-
-public class DockerfileOptions
-{
-    public bool NoFormat { get; set; }
-    public PlatformOptions PlatformOptions { get; set; } = new();
-}
