--- src/Valleysoft.Dredge.Core/ManifestListResolutionException.cs
@@ -1,6 +1,0 @@
-﻿namespace Valleysoft.Dredge.Core;
-
-public class ManifestListResolutionException : Exception
-{
-
-}
