--- src/Valleysoft.Dredge.Core/OsInfo.cs
@@ -1,7 +1,0 @@
-﻿namespace Valleysoft.Dredge.Core;
-
-public class OsInfo
-{
-    public LinuxOsInfo? LinuxOsInfo { get; set; }
-    public WindowsOsInfo? WindowsOsInfo { get; set; }
-}
