--- src/Valleysoft.Dredge.Core/PlatformOptions.cs
@@ -1,8 +1,0 @@
-﻿namespace Valleysoft.Dredge.Core;
-
-public class PlatformOptions
-{
-    public string? Os { get; set; }
-    public string? OsVersion { get; set; }
-    public string? Architecture { get; set; }
-}
