--- src/Valleysoft.Dredge.Core/Valleysoft.Dredge.Core.csproj
@@ -1,15 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
-    <ImplicitUsings>enable</ImplicitUsings>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="SharpZipLib" Version="1.4.1" />
-    <PackageReference Include="Valleysoft.DockerRegistryClient" Version="4.0.0" />
-    <PackageReference Include="Valleysoft.DockerfileModel" Version="1.1.0" />
-  </ItemGroup>
-
-</Project>
