+++ src/Valleysoft.Dredge.Tests/Valleysoft.Dredge.Tests.csproj
@@ -23,7 +23,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Valleysoft.Dredge.Core\Valleysoft.Dredge.Core.csproj" />
     <ProjectReference Include="..\Valleysoft.Dredge\Valleysoft.Dredge.csproj" />
   </ItemGroup>
