+++ src/Valleysoft.Dredge/AppSettings.cs
@@ -1,12 +1,21 @@
 ﻿using Newtonsoft.Json;
-using Valleysoft.Dredge.Core;
 
 namespace Valleysoft.Dredge;
 
-internal class AppSettingsHelper
+internal class AppSettings
 {
     public static readonly string SettingsPath =
         Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData), "Valleysoft.Dredge", "settings.json");
+
+    public const string FileCompareToolName = "fileCompareTool";
+
+    [JsonProperty(FileCompareToolName)]
+    public FileCompareToolSettings FileCompareTool { get; set; } = new();
+
+    [JsonProperty("platform")]
+    public PlatformSettings Platform { get; set; } = new();
+
+    private AppSettings() {}
 
     public static AppSettings Load()
     {
@@ -31,9 +40,30 @@ internal class AppSettingsHelper
         }
     }
 
-    public static void Save(AppSettings appSettings)
+    public void Save()
     {
-        string settingsStr = JsonConvert.SerializeObject(appSettings, Formatting.Indented);
+        string settingsStr = JsonConvert.SerializeObject(this, Formatting.Indented);
         File.WriteAllText(SettingsPath, settingsStr);
     }
+}
+
+internal class FileCompareToolSettings
+{
+    [JsonProperty("exePath")]
+    public string ExePath { get; set; } = string.Empty;
+
+    [JsonProperty("args")]
+    public string Args { get; set; } = string.Empty;
+}
+
+internal class PlatformSettings
+{
+    [JsonProperty("os")]
+    public string Os { get; set; } = string.Empty;
+
+    [JsonProperty("osVersion")]
+    public string OsVersion { get; set; } = string.Empty;
+
+    [JsonProperty("arch")]
+    public string Architecture { get; set; } = string.Empty;
 }
