+++ src/Valleysoft.Dredge/Commands/Image/OsCommand.cs
@@ -1,5 +1,10 @@
-﻿using Newtonsoft.Json;
-using Valleysoft.Dredge.Core;
+﻿using ICSharpCode.SharpZipLib.Tar;
+using Newtonsoft.Json;
+using System.IO.Compression;
+using System.Text;
+using Valleysoft.DockerRegistryClient;
+using Valleysoft.DockerRegistryClient.Models;
+using ImageConfig = Valleysoft.DockerRegistryClient.Models.Image;
 
 namespace Valleysoft.Dredge.Commands.Image;
 
@@ -15,19 +20,109 @@ public class OsCommand : RegistryCommandBase<OsOptions>
         ImageName imageName = ImageName.Parse(Options.Image);
         return CommandHelper.ExecuteCommandAsync(imageName.Registry, async () =>
         {
-            OsInfo osInfo = await OsHelper.GetOsInfoAsync(Options.Image, DockerRegistryClientFactory, AppSettingsHelper.Load(), Options.ToPlatformOptions());
+            using IDockerRegistryClient client = await DockerRegistryClientFactory.GetClientAsync(imageName.Registry);
+            DockerManifestV2 manifest = (await ManifestHelper.GetResolvedManifestAsync(client, imageName, Options)).Manifest;
 
-            if (osInfo.WindowsOsInfo is null && osInfo.LinuxOsInfo is null)
+            string? configDigest = manifest.Config?.Digest;
+            if (configDigest is null)
+            {
+                throw new NotSupportedException($"Could not resolve the image config digest of '{Options.Image}'.");
+            }
+
+            ImageConfig imageConfig = await client.Blobs.GetImageAsync(imageName.Repo, configDigest);
+
+            ManifestLayer baseLayer = manifest.Layers.First();
+            if (baseLayer.Digest is null)
+            {
+                throw new Exception($"No digest was found for the base layer of '{Options.Image}'.");
+            }
+
+            object? osInfo;
+            if (imageConfig.Os.Equals("windows", StringComparison.OrdinalIgnoreCase))
+            {
+                var windowsOsInfo = await GetWindowsOsInfoAsync(imageConfig, baseLayer.Digest, DockerRegistryClientFactory);
+                osInfo = windowsOsInfo?.Info;
+            }
+            else
+            {
+                osInfo = await GetLinuxOsInfoAsync(client, imageName, baseLayer.Digest);
+            }
+
+            if (osInfo is null)
             {
                 throw new Exception("Unable to derive OS information from the image.");
             }
 
-            string output = JsonConvert.SerializeObject((object?)osInfo.WindowsOsInfo ?? osInfo.LinuxOsInfo, new JsonSerializerSettings
+            string output = JsonConvert.SerializeObject(osInfo, new JsonSerializerSettings
             {
                 NullValueHandling = NullValueHandling.Ignore,
                 Formatting = Formatting.Indented
             });
             Console.Out.WriteLine(output);
         });
+    }
+
+    private static async Task<LinuxOsInfo?> GetLinuxOsInfoAsync(IDockerRegistryClient client, ImageName imageName, string baseLayerDigest)
+    {
+        using Stream blobStream = await client.Blobs.GetAsync(imageName.Repo, baseLayerDigest);
+        using GZipStream gZipStream = new(blobStream, CompressionMode.Decompress);
+
+        // Can't use System.Formats.Tar.TarReader because it fails to read certain types of tarballs:
+        // https://github.com/dotnet/runtime/issues/74316#issuecomment-1312227247
+
+        using TarInputStream tarStream = new(gZipStream, Encoding.UTF8);
+        TarEntry? entry = null;
+        do
+        {
+            entry = tarStream.GetNextEntry();
+
+            // Look for the os-release file (skip symlinks)
+            if (entry is not null &&
+                entry.Size > 0 &&
+                (entry.Name == "etc/os-release" || entry.Name == "usr/lib/os-release"))
+            {
+                using MemoryStream memStream = new();
+                tarStream.CopyEntryContents(memStream);
+                memStream.Position = 0;
+                using StreamReader reader = new(memStream);
+                string content = await reader.ReadToEndAsync();
+                return LinuxOsInfo.Parse(content);
+            }
+        } while (entry is not null);
+
+        return null;
+    }
+
+    internal static async Task<(WindowsOsInfo Info, string Repo)?> GetWindowsOsInfoAsync(ImageConfig imageConfig, string baseLayerDigest,
+        IDockerRegistryClientFactory dockerRegistryClientFactory)
+    {
+        const string NanoServerRepo = "windows/nanoserver";
+        const string ServerCoreRepo = "windows/servercore";
+        const string ServerRepo = "windows/server";
+        const string WindowsRepo = "windows";
+
+        using IDockerRegistryClient mcrClient =
+            await dockerRegistryClientFactory.GetClientAsync(RegistryHelper.McrRegistry);
+
+        if (await mcrClient.Blobs.ExistsAsync(NanoServerRepo, baseLayerDigest))
+        {
+            return (new(WindowsType.NanoServer, imageConfig.OsVersion), NanoServerRepo);
+        }
+        else if (await mcrClient.Blobs.ExistsAsync(ServerCoreRepo, baseLayerDigest))
+        {
+            return (new(WindowsType.ServerCore, imageConfig.OsVersion), ServerCoreRepo);
+        }
+        else if (await mcrClient.Blobs.ExistsAsync(ServerRepo, baseLayerDigest))
+        {
+            return (new(WindowsType.Server, imageConfig.OsVersion), ServerRepo);
+        }
+        else if (await mcrClient.Blobs.ExistsAsync(WindowsRepo, baseLayerDigest))
+        {
+            return (new(WindowsType.Windows, imageConfig.OsVersion), WindowsRepo);
+        }
+        else
+        {
+            return null;
+        }
     }
 }
