+++ src/Valleysoft.Dredge/Commands/Manifest/DigestCommand.cs
@@ -1,5 +1,4 @@
 ﻿using Valleysoft.DockerRegistryClient;
-using Valleysoft.Dredge.Core;
 
 namespace Valleysoft.Dredge.Commands.Manifest;
