+++ src/Valleysoft.Dredge/Commands/Manifest/GetCommand.cs
@@ -1,7 +1,6 @@
 ﻿using Newtonsoft.Json;
 using Valleysoft.DockerRegistryClient;
 using Valleysoft.DockerRegistryClient.Models;
-using Valleysoft.Dredge.Core;
 
 namespace Valleysoft.Dredge.Commands.Manifest;
