+++ src/Valleysoft.Dredge/Commands/Manifest/ResolveCommand.cs
@@ -1,5 +1,4 @@
 ﻿using Valleysoft.DockerRegistryClient.Models;
-using Valleysoft.Dredge.Core;
 
 namespace Valleysoft.Dredge.Commands.Manifest;
 
@@ -16,7 +15,7 @@ public class ResolveCommand : RegistryCommandBase<SetOptions>
         return CommandHelper.ExecuteCommandAsync(imageName.Registry, async () =>
         {
             using IDockerRegistryClient client = await DockerRegistryClientFactory.GetClientAsync(imageName.Registry);
-            ManifestInfo manifestInfo = (await ManifestHelper.GetResolvedManifestAsync(client, imageName, AppSettingsHelper.Load(), Options.ToPlatformOptions())).ManifestInfo;
+            ManifestInfo manifestInfo = (await ManifestHelper.GetResolvedManifestAsync(client, imageName, Options)).ManifestInfo;
             ImageName fullyQualifiedDigest = new(imageName.Registry, imageName.Repo, tag: null, manifestInfo.DockerContentDigest);
 
             Console.Out.WriteLine(fullyQualifiedDigest.ToString());
