+++ src/Valleysoft.Dredge/Commands/PlatformOptionsBase.cs
@@ -1,5 +1,4 @@
 ﻿using System.CommandLine;
-using Valleysoft.Dredge.Core;
 
 namespace Valleysoft.Dredge.Commands;
 
@@ -30,12 +29,4 @@ public class PlatformOptionsBase : OptionsBase
         OsVersion = GetValue(osVersionOpt);
         Architecture = GetValue(archOpt);
     }
-
-    public PlatformOptions ToPlatformOptions() =>
-        new PlatformOptions
-        {
-            Os = this.Os,
-            OsVersion = this.OsVersion,
-            Architecture = this.Architecture
-        };
 }
