+++ src/Valleysoft.Dredge/Commands/Settings/SetCommand.cs
@@ -1,6 +1,5 @@
 ﻿using Newtonsoft.Json;
 using System.Reflection;
-using Valleysoft.Dredge.Core;
 
 namespace Valleysoft.Dredge.Commands.Settings;
 
@@ -13,12 +12,12 @@ internal partial class SetCommand : CommandWithOptions<SetOptions>
 
     protected override Task ExecuteAsync()
     {
-        AppSettings settings = AppSettingsHelper.Load();
+        AppSettings settings = AppSettings.Load();
 
         Queue<string> names = new(Options.Name.Split('.'));
 
         SetSettingProperty(settings, names, Options.Value);
-        AppSettingsHelper.Save(settings);
+        settings.Save();
         return Task.CompletedTask;
     }
