+++ src/Valleysoft.Dredge/Commands/Settings/SettingsCommand.cs
@@ -1,5 +1,4 @@
 ﻿using System.CommandLine;
-using Valleysoft.Dredge.Core;
 
 namespace Valleysoft.Dredge.Commands.Settings;
