+++ src/Valleysoft.Dredge/DockerRegistryClientFactory.cs
@@ -1,7 +1,6 @@
 ﻿using Microsoft.Rest;
 using Valleysoft.DockerCredsProvider;
 using Valleysoft.DockerRegistryClient;
-using Valleysoft.Dredge.Core;
 
 namespace Valleysoft.Dredge;
 
@@ -57,7 +56,7 @@ internal class DockerRegistryClientFactory : IDockerRegistryClientFactory
         return new DockerRegistryClientWrapper(CreateClient(registry, clientCreds));
     }
 
-    private static RegistryClient CreateClient(string? registry, ServiceClientCredentials? clientCreds = null)
+    private RegistryClient CreateClient(string? registry, ServiceClientCredentials? clientCreds = null)
     {
         RegistryClient client = new(DockerHubHelper.GetApiRegistry(registry), clientCreds);
         client.HttpClient.Timeout = new TimeSpan(0, 30, 0);
