+++ src/Valleysoft.Dredge/DockerRegistryClientWrapper.cs
@@ -1,5 +1,4 @@
 ﻿using Valleysoft.DockerRegistryClient;
-using Valleysoft.Dredge.Core;
 
 namespace Valleysoft.Dredge;
