+++ src/Valleysoft.Dredge/IDockerRegistryClientFactory.cs
@@ -1,4 +1,4 @@
-﻿namespace Valleysoft.Dredge.Core;
+﻿namespace Valleysoft.Dredge;
 
 public interface IDockerRegistryClientFactory
 {
